finance-agent/finance-agent/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
matplotlib==3.10.0
reportlab==4.2.5
python-pptx==1.0.2
openpyxl==3.1.5
